Remove duy's version of tests, all old tests pass
