+++ wrap/python/gtsam_py/tests/test_PriorFactor.py
@@ -1,9 +1,23 @@
+"""
+GTSAM Copyright 2010-2019, Georgia Tech Research Corporation,
+Atlanta, Georgia 30332-0415
+All Rights Reserved
+
+See LICENSE for the license information
+
+PriorFactor unit tests.
+Author: Frank Dellaert & Duy Nguyen Ta (Python)
+"""
 import unittest
-import gtsam
+
 import numpy as np
 
+import gtsam
+from gtsam.utils.test_case import GtsamTestCase
 
-class TestPriorFactor(unittest.TestCase):
+
+class TestPriorFactor(GtsamTestCase):
+
     def test_PriorFactor(self):
         values = gtsam.Values()
 
@@ -15,12 +29,11 @@ class TestPriorFactor(unittest.TestCase):
         self.assertEqual(factor.error(values), 0)
 
         key = 3
-        priorVector = np.array([0.0, 0.0, 0.0, 0.0])
-        model = gtsam.noiseModel.Unit.Create(4)
+        priorVector = np.array([0., 0., 0.])
+        model = gtsam.noiseModel.Unit.Create(3)
         factor = gtsam.PriorFactorVector(key, priorVector, model)
         values.insert(key, priorVector)
         self.assertEqual(factor.error(values), 0)
-
 
 if __name__ == "__main__":
     unittest.main()
