+++ wrap/python/gtsam_py/tests/test_SFMExample.py
@@ -1,18 +1,31 @@
+"""
+GTSAM Copyright 2010-2019, Georgia Tech Research Corporation,
+Atlanta, Georgia 30332-0415
+All Rights Reserved
+
+See LICENSE for the license information
+
+SFM unit tests.
+Author: Frank Dellaert & Duy Nguyen Ta (Python)
+"""
 import unittest
+
 import numpy as np
 
 import gtsam
+import gtsam.utils.visual_data_generator as generator
 from gtsam import symbol
-import gtsam.tests.visual_data_generator as datagen
+from gtsam.utils.test_case import GtsamTestCase
 
 
-class TestSFMExample(unittest.TestCase):
+class TestSFMExample(GtsamTestCase):
+
     def test_SFMExample(self):
-        options = datagen.Options()
+        options = generator.Options()
         options.triangle = False
         options.nrCameras = 10
 
-        [data, truth] = datagen.generate_data(options)
+        [data, truth] = generator.generate_data(options)
 
         measurementNoiseSigma = 1.0
         pointNoiseSigma = 0.1
@@ -21,34 +34,20 @@ class TestSFMExample(unittest.TestCase):
         graph = gtsam.NonlinearFactorGraph()
 
         # Add factors for all measurements
-        measurementNoise = gtsam.noiseModel.Isotropic.Sigma(
-            2, measurementNoiseSigma
-        )
+        measurementNoise = gtsam.noiseModel.Isotropic.Sigma(2, measurementNoiseSigma)
         for i in range(len(data.Z)):
             for k in range(len(data.Z[i])):
                 j = data.J[i][k]
-                graph.add(
-                    gtsam.GenericProjectionFactorCal3_S2(
-                        data.Z[i][k],
-                        measurementNoise,
-                        symbol('x', i),
-                        symbol('p', j),
-                        data.K,
-                    )
-                )
+                graph.add(gtsam.GenericProjectionFactorCal3_S2(
+                    data.Z[i][k], measurementNoise,
+                    symbol('x', i), symbol('p', j), data.K))
 
         posePriorNoise = gtsam.noiseModel.Diagonal.Sigmas(poseNoiseSigmas)
-        graph.add(
-            gtsam.PriorFactorPose3(
-                symbol('x', 0), truth.cameras[0].pose(), posePriorNoise
-            )
-        )
+        graph.add(gtsam.PriorFactorPose3(symbol('x', 0),
+                                   truth.cameras[0].pose(), posePriorNoise))
         pointPriorNoise = gtsam.noiseModel.Isotropic.Sigma(3, pointNoiseSigma)
-        graph.add(
-            gtsam.PriorFactorPoint3(
-                symbol('p', 0), truth.points[0], pointPriorNoise
-            )
-        )
+        graph.add(gtsam.PriorFactorPoint3(symbol('p', 0),
+                                    truth.points[0], pointPriorNoise))
 
         # Initial estimate
         initialEstimate = gtsam.Values()
@@ -73,12 +72,11 @@ class TestSFMExample(unittest.TestCase):
         # Check optimized results, should be equal to ground truth
         for i in range(len(truth.cameras)):
             pose_i = result.atPose3(symbol('x', i))
-            self.assertTrue(pose_i.equals(truth.cameras[i].pose(), 1e-5))
+            self.gtsamAssertEquals(pose_i, truth.cameras[i].pose(), 1e-5)
 
         for j in range(len(truth.points)):
             point_j = result.atPoint3(symbol('p', j))
-            np.testing.assert_almost_equal(point_j, truth.points[j], 1e-5)
-
+            self.gtsamAssertEquals(point_j, truth.points[j], 1e-5)
 
 if __name__ == "__main__":
     unittest.main()
