+++ wrap/python/gtsam_py/tests/test_SO4.py
