+++ wrap/python/gtsam_py/tests/test_initialize_pose3.py
